requirements.txt
cassandra-driver==3.25.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
geomet==0.2.1.post1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pip==21.0.1
python-dotenv==0.16.0
requests==2.25.1
setuptools==54.1.2
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
wheel==0.36.2